requirements.txt
asgiref==3.9.0
Django==5.2.4
django-environ==0.12.0
django-tenants==3.8.0
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
